fix most exampels
