+++ examples/distributed_authority/src/server.rs
@@ -13,7 +13,6 @@ use bevy::prelude::*;
 use bevy::utils::HashMap;
 use lightyear::prelude::server::*;
 use lightyear::prelude::*;
-use lightyear::server::replication::send::ReplicationTarget;
 use std::sync::Arc;
 use std::time::Duration;
