+++ examples/interest_management/Cargo.toml
@@ -13,7 +13,6 @@ publish = false
 
 [features]
 default = [
-    "lightyear_examples_common/steam",
     "lightyear_examples_common/websocket"
 ]
 client = [
