+++ examples/interest_management/src/shared.rs
@@ -1,6 +1,5 @@
 use bevy::color::palettes::css::GREEN;
 use bevy::prelude::*;
-use bevy::render::RenderPlugin;
 use bevy::utils::Duration;
 use leafwing_input_manager::action_state::ActionState;
 use std::ops::Deref;
@@ -16,15 +15,7 @@ pub struct SharedPlugin;
 impl Plugin for SharedPlugin {
     fn build(&self, app: &mut App) {
         app.add_plugins(ProtocolPlugin);
-        if app.is_plugin_added::<RenderPlugin>() {
-            app.add_systems(Startup, init);
-            app.add_systems(Update, (draw_boxes, draw_circles));
-        }
     }
-}
-
-fn init(mut commands: Commands) {
-    commands.spawn(Camera2d::default());
 }
 
 // This system defines how we update the player's positions when we receive an input
@@ -41,29 +32,6 @@ pub(crate) fn shared_movement_behaviour(mut position: Mut<Position>, input: &Act
     }
     if input.pressed(&Inputs::Right) {
         position.x += MOVE_SPEED;
-    }
-}
-
-/// System that draws the boxed of the player positions.
-/// The components should be replicated from the server to the client
-/// This time we will only draw the predicted/interpolated entities
-pub(crate) fn draw_boxes(
-    mut gizmos: Gizmos,
-    players: Query<(&Position, &PlayerColor), Without<Confirmed>>,
-) {
-    for (position, color) in &players {
-        gizmos.rect(
-            Isometry3d::from_translation(Vec3::new(position.x, position.y, 0.0)),
-            Vec2::ONE * 50.0,
-            color.0,
-        );
-    }
-}
-
-/// System that draws circles
-pub(crate) fn draw_circles(mut gizmos: Gizmos, circles: Query<&Position, With<CircleMarker>>) {
-    for position in &circles {
-        gizmos.circle_2d(*position.deref(), 1.0, GREEN);
     }
 }
