+++ examples/lobby/src/renderer.rs
@@ -1,13 +1,18 @@
+use crate::protocol::*;
 use bevy::prelude::*;
 use bevy::render::RenderPlugin;
-
-use crate::protocol::*;
+use bevy_egui::EguiPlugin;
+use lightyear::client::interpolation::Interpolated;
+use lightyear::client::prediction::Predicted;
 
 #[derive(Clone)]
 pub struct ExampleRendererPlugin;
 
 impl Plugin for ExampleRendererPlugin {
     fn build(&self, app: &mut App) {
+        if !app.is_plugin_added::<EguiPlugin>() {
+            app.add_plugins(EguiPlugin);
+        }
         app.add_systems(Startup, init);
         app.add_systems(Update, draw_boxes);
     }
@@ -19,10 +24,13 @@ fn init(mut commands: Commands) {
 
 /// System that draws the boxes of the player positions.
 /// The components should be replicated from the server to the client
-pub(crate) fn draw_boxes(mut gizmos: Gizmos, players: Query<(&PlayerPosition, &PlayerColor)>) {
+pub(crate) fn draw_boxes(
+    mut gizmos: Gizmos,
+    players: Query<(&PlayerPosition, &PlayerColor), Or<(With<Predicted>, With<Interpolated>)>>,
+) {
     for (position, color) in &players {
-        gizmos.rect_2d(
-            Isometry2d::from_translation(position.0),
+        gizmos.rect(
+            Isometry3d::from_translation(Vec3::new(position.x, position.y, 0.0)),
             Vec2::ONE * 50.0,
             color.0,
         );
