+++ examples/priority/Cargo.toml
@@ -14,7 +14,6 @@ publish = false
 
 [features]
 default = [
-    "lightyear_examples_common/steam",
     "lightyear_examples_common/websocket"
 ]
 client = [
