+++ examples/replication_groups/src/client.rs
@@ -12,7 +12,6 @@ pub struct ExampleClientPlugin;
 impl Plugin for ExampleClientPlugin {
     fn build(&self, app: &mut App) {
         app.add_systems(Startup, init);
-        app.add_systems(PreUpdate, handle_connection.after(MainSet::Receive));
         // app.add_systems(
         //     PostUpdate,
         //     debug_interpolate
@@ -42,22 +41,6 @@ impl Plugin for ExampleClientPlugin {
 // Startup system for the client
 pub(crate) fn init(mut commands: Commands) {
     commands.connect_client();
-}
-
-/// Listen for events to know when the client is connected, and spawn a text entity
-/// to display the client id
-pub(crate) fn handle_connection(
-    mut commands: Commands,
-    mut connection_event: EventReader<ConnectEvent>,
-) {
-    for event in connection_event.read() {
-        let client_id = event.client_id();
-        commands.spawn((
-            Text(format!("Client {}", client_id)),
-            TextFont::from_font_size(30.0),
-            TextColor(Color::WHITE),
-        ));
-    }
 }
 
 // System that reads from peripherals and adds inputs to the buffer
