+++ examples/simple_box/src/client.rs
@@ -24,8 +24,6 @@ pub struct ExampleClientPlugin;
 
 impl Plugin for ExampleClientPlugin {
     fn build(&self, app: &mut App) {
-        // app.add_systems(Startup, spawn_connect_button);
-
         // Inputs have to be buffered in the FixedPreUpdate schedule
         app.add_systems(
             FixedPreUpdate,
