+++ examples/simple_box/src/server.rs
@@ -11,7 +11,6 @@ use bevy::prelude::*;
 use bevy::utils::HashMap;
 use lightyear::prelude::server::*;
 use lightyear::prelude::*;
-use lightyear::server::replication::send::ReplicationTarget;
 use std::sync::Arc;
 
 use crate::protocol::*;
